% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{from01}
\alias{from01}
\title{Helper function to scale from [0,1] to [min,max]}
\usage{
from01(X, inp)
}
\arguments{
\item{X}{values to scale}

\item{inp}{list containing 'min' and 'max' values}
}
\value{
X scaled in [inp$min, inp$max]
}
\description{
Helper function to scale from [0,1] to [min,max]
}
\examples{
from01(data.frame(x=matrix(runif(10))),list(x=list(min=10,max=20)))
}
