% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{max_input}
\alias{max_input}
\title{Helper function to get $max from 'input' list}
\usage{
max_input(inp)
}
\arguments{
\item{inp}{lst of objects containing 'max' field (as list)}
}
\value{
array of inp$...$max values
}
\description{
Helper function to get $max from 'input' list
}
\examples{
max_input(list(x1=list(min=0,max=1),x2=list(min=2,max=3)))
}
