% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{min_input}
\alias{min_input}
\title{Helper function to get $min from 'input' list}
\usage{
min_input(inp)
}
\arguments{
\item{inp}{lst of objects containing 'min' field (as list)}
}
\value{
array of inp$...$min values
}
\description{
Helper function to get $min from 'input' list
}
\examples{
min_input(list(x1=list(min=0,max=1),x2=list(min=2,max=3)))
}
