% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{parse.algorithm}
\alias{parse.algorithm}
\title{Parse algorithm file and returns its (header) indos and methods}
\usage{
parse.algorithm(file)
}
\arguments{
\item{file}{Template algorithm file to parse}
}
\value{
list of header infos and environment containing methods <constructor>,getInitialDesign,getNextDesign,displayResults
}
\description{
Parse algorithm file and returns its (header) indos and methods
}
\examples{
parse.algorithm(system.file("Brent.R", package="templr"))
}
