% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{read.algorithm}
\alias{read.algorithm}
\title{Read algorithm file and returns one header info}
\usage{
read.algorithm(file, info = "help")
}
\arguments{
\item{file}{Template algorithm file to read}

\item{info}{header info to return}
}
\value{
list of header infos
}
\description{
Read algorithm file and returns one header info
}
\examples{
read.algorithm(system.file("Brent.R", package="templr"),"help")
}
