% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{to01}
\alias{to01}
\title{Helper function to scale from [min,max] to [0,1]}
\usage{
to01(X, inp)
}
\arguments{
\item{X}{values to scale}

\item{inp}{list containing 'min' and 'max' values}
}
\value{
X scaled in [0,1]
}
\description{
Helper function to scale from [min,max] to [0,1]
}
\examples{
to01(10+10*data.frame(x=matrix(runif(10))),list(x=list(min=10,max=20)))
}
