% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asktell.R
\name{ask_dX}
\alias{ask_dX}
\title{ask&tell component function to 'ask' where objective function gradient evaluation is required.}
\usage{
ask_dX(
  id = 0,
  dX.tmp = "dX.todo",
  tmp_path = file.path(tempdir(), "..", "asktell.tmp"),
  sleep_step = 0.1,
  sleep_init = 0,
  timeout = 360000,
  trace = function(...) cat(paste0(..., "\\n")),
  clean = TRUE
)
}
\arguments{
\item{id}{unique identifier for this asktell loop (default: "0")}

\item{dX.tmp}{temporary "X" values file (default: "dX.todo")}

\item{tmp_path}{temporary directory to store X.tmp & Y.tmp (default: 'tempdir()/../asktell.tmp')}

\item{sleep_step}{delay between checking X.tmp and Y.tmp (default: 0.1 sec.)}

\item{sleep_init}{initial delay before checking X.tmp and Y.tmp  (default: 0 sec.)}

\item{timeout}{maximum delay before breaking loop if X.tmp or Y.tmp doesn't appear (default: 36000 sec. = 10 min.) .}

\item{trace}{function to display asktell loop status (default : 'cat')}

\item{clean}{should we cleanup temporary files after reading ? (default: TRUE)}
}
\value{
input values of objective function to compute externally
}
\description{
ask&tell component function to 'ask' where objective function gradient evaluation is required.
}
\details{
'ask&tell' injection loop to call an external objective function within an inline algorithm (like optim(...))
Main idea: pass 'ask_Y' as objectve function argument of algorithm, which will wait until you call 'tell_Y' in another R process.
In this secondary process, you can read what X is called using 'ask_X', and when you know what values returns from the external objective, 
just call 'tell_Y' to give it.
}
\examples{
\dontrun{ ### Assumes you can use two independant R sessions
## In main R session
  ask_dY(x=123)
## In another R session
  ask_dX() # returns 123
  tell_dY(y=456)
## Then ask_dY in main R session returns with value '456'
}
}
\author{
Y. Richet, discussions with D. Sinoquet. Async IO principle was defined by G. Pujol.
}
