% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Dependencies loader, supports many protocols like github:, gitlab:, ... using remotes::instal_...
Will create a local '.lib' directory to store packages installed}
\usage{
import(..., trace = function(...) cat(paste0(..., "\\n")))
}
\arguments{
\item{...}{dependencies/libraries/packages to load}

\item{trace}{display info}
}
\description{
Dependencies loader, supports many protocols like github:, gitlab:, ... using remotes::instal_...
Will create a local '.lib' directory to store packages installed
}
\examples{
\dontrun{
  options(data.frame(repos="http://cran.irsn.fr")); import('VGAM')
}
}
