% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{list.results}
\alias{list.results}
\title{Parse algorithm string result in R list}
\usage{
list.results(result)
}
\arguments{
\item{result}{templated algorithm result string}
}
\value{
list of string parsed: extract XML or JSON content
}
\description{
Parse algorithm string result in R list
}
